02_웹_0912/FastAPI_part1_old/requirements.txt
cryptography==41.0.3
fastapi==0.103.1
LibRecommender==1.3.0
mangum==0.17.0
pandas==2.1.0
python-multipart==0.0.6
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
